% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alle_logit.R
\name{alle_logit}
\alias{alle_logit}
\title{Almost Unbiased Liu Logistic Estimator (ALLE)}
\usage{
alle_logit(X, y, d = 0.5, standardize = FALSE, tol = 1e-06, max_iter = 50)
}
\arguments{
\item{X}{Matrix or data frame of predictors.}

\item{y}{Binary response vector (0/1).}

\item{d}{Liu shrinkage parameter, numeric in [0, 1].}

\item{tol}{Convergence tolerance for IRLS (default = 1e-6).}

\item{max_iter}{Maximum number of IRLS iterations (default = 50).}
}
\value{
An object of class "alle_logit" similar to glm output.
}
\description{
Fits an Almost Unbiased Liu Logistic Regression model using IRLS.
This estimator reduces the bias of the standard Liu logistic estimator.
}
\details{
The estimator is defined as:
\deqn{\hat{\beta}_{ALLE} = [I - (1 + d)(X'WX + I)^{-1}] \hat{\beta}_{MLE}}
}
