% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_logit.R
\name{liu_logit}
\alias{liu_logit}
\title{Liu Logistic Regression}
\usage{
liu_logit(X, y, d = 0, standardize = FALSE, tol = 1e-06, max_iter = 50)
}
\arguments{
\item{X}{Matrix or data frame of predictors (including intercept or not).}

\item{y}{Binary response vector (0/1).}

\item{d}{Liu biasing parameter (0 <= d < 1).}

\item{tol}{Convergence tolerance (default = 1e-6).}

\item{max_iter}{Maximum number of iterations (default = 50).}
}
\value{
Object of class "liu_logit" containing coefficients, deviance, etc.
}
\description{
Fits a Liu-penalized logistic regression model.
Returns an object of class "liu_logit" similar to glm output.
}
