% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmrt_logit.R
\name{lmrt_logit}
\alias{lmrt_logit}
\title{Logistic Modified Ridge Estimator (LMRT)}
\usage{
lmrt_logit(X, y, k = 1, d = 0.5)
}
\arguments{
\item{X}{Matrix or data frame of predictors (without intercept)}

\item{y}{Binary response vector (0 or 1)}

\item{k}{Ridge-type biasing parameter (k >= 0)}

\item{d}{Additional bias parameter (0 <= d <= 1)}
}
\value{
An object of class "lmrt_logit"
}
\description{
Fits the Logistic Modified Ridge Estimator (LMRT) using:
\deqn{\hat{\beta}_{MRT} = (X'WX + k(1 + d)I)^{-1}(X'WX)\hat{\beta}_{MLE}}
}
