% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lske_logit.R
\name{lske_logit}
\alias{lske_logit}
\title{Logistic Stein K-Liu Estimator (LSKE)}
\usage{
lske_logit(X, y, k = 1)
}
\arguments{
\item{X}{Matrix or data frame of predictors (without intercept).}

\item{y}{Binary response vector (0/1).}

\item{k}{Biasing parameter (k >= 0).}
}
\value{
An object of class "lske_logit" similar to glm output.
}
\description{
Implements the Logistic Stein K-Liu Estimator (LSKE) as defined in:
\deqn{β̂_LSKE = c_k K β̂_MLE}
where
\deqn{K = (X'WX + kI)^{-1}(X'WX - kI)}
and
\deqn{c_k = (β̂_MLE'K'β̂_MLE) / (β̂_MLE'K'Kβ̂_MLE + tr(MSE * (K(X'WX)^{-1}K')))}
}
