% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsre_logit.R
\name{lsre_logit}
\alias{lsre_logit}
\title{Logistic Stein Ridge Estimator (LSRE)}
\usage{
lsre_logit(X, y, k = 1)
}
\arguments{
\item{X}{Matrix or data frame of predictors (without intercept).}

\item{y}{Binary response vector (0/1).}

\item{k}{Ridge-type shrinkage parameter (k >= 0).}
}
\value{
An object of class "lsre_logit" similar to glm.
}
\description{
Implements the Logistic Stein Ridge Estimator (LSRE):
\deqn{β̂_LSRE = c_r A β̂_MLE}
where
\deqn{A = (X'WX + kI)^{-1} X'WX}
and
\deqn{c_r = (β̂_MLE'A'Aβ̂_MLE) / (β̂_MLE'A'Aβ̂_MLE + tr(MSE * (A(X'WX)^{-1}A')))}
}
