% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predict}
\alias{predict}
\title{Generic predict function for LogisticEstimators package}
\usage{
predict(object, ...)
}
\arguments{
\item{object}{A fitted model object (e.g., of class "ridge_logit", "liu_logit", etc.)}

\item{...}{Additional arguments passed to specific methods.}
}
\description{
This function dispatches to the appropriate `predict.<estimator>` method
depending on the class of the fitted model.
}
