% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_logit.R
\name{ridge_logit}
\alias{ridge_logit}
\title{Ridge Logistic Regression}
\usage{
ridge_logit(X, y, lambda = 0, standardize = FALSE, tol = 1e-06, max_iter = 50)
}
\arguments{
\item{X}{A matrix or data frame of predictors.}

\item{y}{A binary response vector (0 or 1).}

\item{lambda}{The ridge penalty parameter (default = 1).}

\item{tol}{Convergence tolerance (default = 1e-6).}

\item{maxit}{Maximum number of iterations (default = 50).}
}
\value{
An object of class "ridge_logit" similar to glm output.
}
\description{
Fits a ridge-penalized logistic regression model using Iteratively Reweighted Least Squares (IRLS).
Returns an object of class "ridge_logit" similar to glm output.
}
