% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_logit.R
\name{summary.liu_logit}
\alias{summary.liu_logit}
\title{Summary Method for Liu Logistic Regression}
\usage{
\method{summary}{liu_logit}(object, ...)
}
\description{
Summary Method for Liu Logistic Regression
}
